BUG: Fixed read float MRML attribute macro
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=26937
git-svn-id: http://svn.slicer.org/Slicer4/trunk@26937 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/Core/vtkMRMLNodePropertyMacros.h
@@ -177,7 +177,7 @@
     { \
     vtkVariant variantValue(xmlReadAttValue); \
     bool valid = false; \
-    int scalarValue =  variantValue.ToFloat(&valid); \
+    double scalarValue =  variantValue.ToDouble(&valid); \
     if (valid) \
       { \
       this->Set##propertyName(scalarValue); \
